fix : sound structure
Moved from a VecDeque to a Vec
